hw4-7-more-1

+++ Homework/HW04/hw04-7-Temperator-Convert.cpp
@@ -15,10 +15,10 @@
 */
 #include <stdio.h>
 int main() {
-    int Celsius = 0,Fahrenheit = 0  ;
+    float Celsius = 0,Fahrenheit = 0  ;
     printf("Enter you degree Celsius: ");
-    scanf("%d",&Celsius ) ;
+    scanf("%f",&Celsius ) ;
     Fahrenheit = 1.8 * Celsius + 32  ;
-    printf("%d degree Celsius = %d degree Fahrenheit.\n ", Celsius, Fahrenheit ) ;
+    printf("%0.1f degree Celsius = %0.1f degree Fahrenheit.\n ", Celsius, Fahrenheit ) ;
     return 0 ;
 }
